# backend: removes the map use in ExecutableContext[Instr]